Restructure and unstable features
This commit is too big... It restructures that tests so that there are
now unstable and stable tests. It also adds a new structure that
supports feature gates. The `length` and `range` validators have now
been changed back to their previous design using `a, b` variables
but there is now a feature gate *inclusive_range* that enables the
use of RangeInclusive in them instead.